feat(plugins): Added anonymize_identifiable_data to access_control #201
