+++ ozpcenter/api/listing/tests/test_api_listing.py
@@ -1,8 +1,6 @@
 """
 Tests for listing endpoints
 """
-import json
-
 from rest_framework import status
 from rest_framework.test import APITestCase
 
@@ -87,507 +85,6 @@ class ListingApiTest(APITestCase):
                 raise Exception('status code is not supported')
 
         return response
-
-    def test_search_categories_single_with_space(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?category=Health and Fitness'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data]
-        self.assertTrue('Hatch Latch' in titles)
-        self.assertEqual(len(titles), 20)
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_search_categories_multiple_with_space(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?category=Health and Fitness&category=Communication'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data]
-        self.assertTrue('Air Mail' in titles)
-        self.assertTrue('Bread Basket' in titles)
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_search_text(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?search=air ma'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data]
-        self.assertTrue('Air Mail' in titles)
-        self.assertEqual(len(titles), 10)
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_search_type(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?type=web application'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data]
-        self.assertTrue('Air Mail' in titles)
-        self.assertTrue(len(titles) > 7)
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_search_tags(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?search=tag_1'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data]
-        self.assertTrue('Air Mail 1' in titles)
-        self.assertTrue(len(titles) == 1)
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_search_tags_startwith(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?search=tag_'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data]
-        self.assertTrue('Air Mail 1' in titles)
-        self.assertTrue('Air Mail' in titles)
-        self.assertTrue(len(titles) == 10)
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_search_is_enable(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?search=airmail&type=web application'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        ids = [record.get('id') for record in response.data]
-        self.assertEqual(ids, [1, 12, 23, 34, 45, 56, 67, 78, 89, 100])
-
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-        # Disable one app
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-
-        url = '/api/listing/1/'
-        title = 'airmail_disabled'
-
-        data = {
-            "title": title,
-            "description": "description of app",
-            "launch_url": "http://www.google.com/launch",
-            "version_name": "1.0.0",
-            "unique_name": "org.apps.julia-one",
-            "what_is_new": "nothing is new",
-            "description_short": "a shorter description",
-            "requirements": "None",
-            "is_private": "true",
-            "is_enable": "false",
-            "contacts": [
-                {"email": "a@a.com", "secure_phone": "111-222-3434",
-                 "unsecure_phone": "444-555-4545", "name": "me",
-                 "contact_type": {"name": "Government"}
-                    },
-                {"email": "b@b.com", "secure_phone": "222-222-3333",
-                 "unsecure_phone": "555-555-5555", "name": "you",
-                 "contact_type": {"name": "Military"}
-                    }
-                ],
-            "security_marking": "UNCLASSIFIED",
-            "listing_type": {"title": "web application"},
-            "small_icon": {"id": 1},
-            "large_icon": {"id": 2},
-            "banner_icon": {"id": 3},
-            "large_banner_icon": {"id": 4},
-            "categories": [
-                {"title": "Business"},
-                {"title": "Education"}
-                ],
-            "owners": [
-                {"user": {"username": "wsmith"}},
-                {"user": {"username": "julia"}}
-                ],
-            "tags": [
-                {"name": "demo"},
-                {"name": "map"}
-                ],
-            "intents": [
-                {"action": "/application/json/view"},
-                {"action": "/application/json/edit"}
-                ],
-            "doc_urls": [
-                {"name": "wiki", "url": "http://www.google.com/wiki"},
-                {"name": "guide", "url": "http://www.google.com/guide"}
-                ],
-            "screenshots": [
-                {"small_image": {"id": 1}, "large_image": {"id": 2}},
-                {"small_image": {"id": 3}, "large_image": {"id": 4}}
-                ]
-
-            }
-
-        response = self.client.put(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        # Check
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?search=airmail&type=web application'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        ids = [record.get('id') for record in response.data]
-        self.assertEqual(ids, [12, 23, 34, 45, 56, 67, 78, 89, 100])
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_search_agency(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?agency=Minipax&agency=Miniluv'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data]
-        self.assertTrue('Chatter Box' in titles)
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_search_limit(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?limit=1'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data['results']]
-        self.assertTrue('JotSpot' in titles)
-        self.assertEqual(len(titles), 1)
-        for listing_map in response.data['results']:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_search_offset_limit(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/search/?offset=1&limit=1'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data['results']]
-        self.assertTrue('JotSpot 1' in titles)
-        self.assertEqual(len(titles), 1)
-        for listing_map in response.data['results']:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_self_listing(self):
-        user = generic_model_access.get_profile('julia').user
-        self.client.force_authenticate(user=user)
-        url = '/api/self/listing/'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['title'] for i in response.data]
-        self.assertTrue('Bread Basket' in titles)
-        self.assertTrue('Chatter Box' in titles)
-        self.assertTrue('Air Mail' not in titles)
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
-
-    def test_get_reviews(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        air_mail_id = models.Listing.objects.get(title='Air Mail').id
-        url = '/api/listing/{0!s}/review/'.format(air_mail_id)
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-    def test_get_single_review(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        air_mail_id = models.Listing.objects.get(title='Air Mail').id
-        url = '/api/listing/{0!s}/review/1/'.format(air_mail_id)
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertTrue('rate' in response.data)
-        self.assertTrue('text' in response.data)
-        self.assertTrue('author' in response.data)
-        self.assertTrue('listing' in response.data)
-
-    def test_create_review(self):
-        # create a new review
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        air_mail_id = models.Listing.objects.get(title='Air Mail').id
-        url = '/api/listing/{0!s}/review/'.format(air_mail_id)
-        data = {'rate': 4, 'text': 'winston test review'}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        created_id = response.data['id']
-        new_review = models.Review.objects.get(id=created_id)
-        self.assertEqual(created_id, new_review.id)
-
-        # creating a duplicate review should fail
-        # http://stackoverflow.com/questions/21458387/transactionmanagementerror-you-cant-execute-queries-until-the-end-of-the-atom
-        # try:
-        #     with transaction.atomic():
-        #         response = self.client.post(url, data, format='json')
-        #         # self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        #         self.assertTrue(0, 'Duplicate question allowed.')
-        # except IntegrityError:
-        #     pass
-
-    def test_create_review_not_found(self):
-        # creating a review for an app this user cannot see should fail
-        bread_basket_id = models.Listing.objects.get(title='Bread Basket').id
-        user = generic_model_access.get_profile('rutherford').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listing/{0!s}/review/'.format(bread_basket_id)
-        data = {'rate': 4, 'text': 'rutherford test review'}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_create_review_no_text(self):
-        # create a new review
-        user = generic_model_access.get_profile('julia').user
-        self.client.force_authenticate(user=user)
-        air_mail_id = models.Listing.objects.get(title='Air Mail').id
-        url = '/api/listing/{0!s}/review/'.format(air_mail_id)
-        data = {'rate': 3}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        created_id = response.data['id']
-        new_review = models.Review.objects.get(id=created_id)
-        self.assertEqual(created_id, new_review.id)
-
-    def test_update_review(self):
-        """
-        Also tests the listing/<id>/activity endpoint
-        """
-        # create a new review
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        air_mail_id = models.Listing.objects.get(title='Air Mail').id
-        url = '/api/listing/{0!s}/review/'.format(air_mail_id)
-        data = {'rate': 4, 'text': 'winston test review'}
-        response = self.client.post(url, data, format='json')
-        review_id = response.data['id']
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        # now update it
-        url = '/api/listing/{0!s}/review/{1!s}/'.format(air_mail_id, review_id)
-        data = {'rate': 4, 'text': 'winston test review'}
-        response = self.client.put(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(4, response.data['rate'])
-
-        # test the listing/<id>/activity endpoint
-        url = '/api/listing/{0!s}/activity/'.format(air_mail_id)
-        response = self.client.get(url, format='json')
-        activiy_actions = [i['action'] for i in response.data]
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertTrue(models.ListingActivity.REVIEW_EDITED in activiy_actions)
-
-        # try to edit a review from another user - should fail
-        url = '/api/listing/{0!s}/review/1/'.format(air_mail_id)
-        data = {'rate': 4, 'text': 'winston test review'}
-        response = self.client.put(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_simple_delete_review(self):
-        # create a new review
-        user = generic_model_access.get_profile('rutherford').user
-        self.client.force_authenticate(user=user)
-        air_mail_id = models.Listing.objects.get(title='Air Mail').id
-        url = '/api/listing/{0!s}/review/'.format(air_mail_id)
-        data = {'rate': 4, 'text': 'rutherford test review'}
-        response = self.client.post(url, data, format='json')
-        review_id = response.data['id']
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        # now delete it
-        url = '/api/listing/{0!s}/review/{1!s}/'.format(air_mail_id, review_id)
-        response = self.client.delete(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
-
-        # test the listing/<id>/activity endpoint
-        url = '/api/listing/{0!s}/activity/'.format(air_mail_id)
-        response = self.client.get(url, format='json')
-        activiy_actions = [i['action'] for i in response.data]
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertTrue(models.ListingActivity.REVIEW_DELETED in activiy_actions)
-
-    def test_delete_review(self):
-        # create a new review
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        air_mail_id = models.Listing.objects.get(title='Air Mail').id
-        url = '/api/listing/{0!s}/review/'.format(air_mail_id)
-        data = {'rate': 4, 'text': 'winston test review'}
-        response = self.client.post(url, data, format='json')
-        review_id = response.data['id']
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        # trying to delete it as a different user should fail...
-        url = '/api/listing/{0!s}/review/{1!s}/'.format(air_mail_id, review_id)
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-        response = self.client.delete(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-        # ... unless that user is an org steward or apps mall steward
-        user = generic_model_access.get_profile('julia').user
-        self.client.force_authenticate(user=user)
-        response = self.client.delete(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
-
-    def test_rating_updates(self):
-        """
-        Tests that reviews are updated
-        """
-        title = 'Hatch Latch'
-        # get a listing with no reviews
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        listing = models.Listing.objects.get(title=title)
-        self.assertEqual(listing.avg_rate, 0.0)
-        self.assertEqual(listing.total_votes, 0)
-        self.assertEqual(listing.total_reviews, 0)
-        self.assertEqual(listing.total_rate1, 0)
-        self.assertEqual(listing.total_rate2, 0)
-        self.assertEqual(listing.total_rate3, 0)
-        self.assertEqual(listing.total_rate4, 0)
-        self.assertEqual(listing.total_rate5, 0)
-
-        # add one review
-        url = '/api/listing/{0!s}/review/'.format(listing.id)
-        data = {'rate': 1, 'text': 'winston test review'}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        # check calculations
-        listing = models.Listing.objects.get(title=title)
-        self.assertEqual(listing.avg_rate, 1.0)
-        self.assertEqual(listing.total_votes, 1)
-        self.assertEqual(listing.total_reviews, 1)
-        self.assertEqual(listing.total_rate1, 1)
-        self.assertEqual(listing.total_rate2, 0)
-        self.assertEqual(listing.total_rate3, 0)
-        self.assertEqual(listing.total_rate4, 0)
-        self.assertEqual(listing.total_rate5, 0)
-
-        # add a few more reviews
-        user = generic_model_access.get_profile('julia').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listing/{0!s}/review/'.format(listing.id)
-        data = {'rate': 2, 'text': 'julia test review'}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        user = generic_model_access.get_profile('charrington').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listing/{0!s}/review/'.format(listing.id)
-        data = {'rate': 3, 'text': 'charrington test review'}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listing/{0!s}/review/'.format(listing.id)
-        data = {'rate': 4, 'text': 'jones test review'}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        user = generic_model_access.get_profile('rutherford').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listing/{0!s}/review/'.format(listing.id)
-        data = {'rate': 5, 'text': 'rutherford test review'}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-
-        user = generic_model_access.get_profile('syme').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listing/{0!s}/review/'.format(listing.id)
-        data = {'rate': 5, 'text': 'syme test review'}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        syme_review_id = response.data['id']
-
-        # and one without a review
-        user = generic_model_access.get_profile('tparsons').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listing/{0!s}/review/'.format(listing.id)
-        data = {'rate': 4}
-        response = self.client.post(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
-        tparsons_review_id = response.data['id']
-
-        # check calculations
-        listing = models.Listing.objects.get(title=title)
-        # (2*5 + 2*4 + 1*3 + 1*2 + 1*1)/7 = (24)/7 = 3.429
-        self.assertEqual(listing.avg_rate, 3.4)
-        self.assertEqual(listing.total_votes, 7)
-        self.assertEqual(listing.total_reviews, 6)
-        self.assertEqual(listing.total_rate1, 1)
-        self.assertEqual(listing.total_rate2, 1)
-        self.assertEqual(listing.total_rate3, 1)
-        self.assertEqual(listing.total_rate4, 2)
-        self.assertEqual(listing.total_rate5, 2)
-
-        # update an existing review
-        user = generic_model_access.get_profile('tparsons').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listing/{0!s}/review/{1!s}/'.format(listing.id,
-            tparsons_review_id)
-        data = {'rate': 2}
-        response = self.client.put(url, data, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        # check calculations
-        listing = models.Listing.objects.get(title=title)
-        # (2*5 + 1*4 + 1*3 + 2*2 + 1*1)/7 = (22)/7 = 3.14
-        self.assertEqual(listing.avg_rate, 3.1)
-        self.assertEqual(listing.total_votes, 7)
-        self.assertEqual(listing.total_reviews, 6)
-        self.assertEqual(listing.total_rate1, 1)
-        self.assertEqual(listing.total_rate2, 2)
-        self.assertEqual(listing.total_rate3, 1)
-        self.assertEqual(listing.total_rate4, 1)
-        self.assertEqual(listing.total_rate5, 2)
-
-        # delete an existing review
-        user = generic_model_access.get_profile('syme').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listing/{0!s}/review/{1!s}/'.format(listing.id, syme_review_id)
-        response = self.client.delete(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
-
-        # check calculations
-        listing = models.Listing.objects.get(title=title)
-        # (1*5 + 1*4 + 1*3 + 2*2 + 1*1)/6 = (17)/6 = 2.83
-        self.assertEqual(listing.avg_rate, 2.8)
-        self.assertEqual(listing.total_votes, 6)
-        self.assertEqual(listing.total_reviews, 5)
-        self.assertEqual(listing.total_rate1, 1)
-        self.assertEqual(listing.total_rate2, 2)
-        self.assertEqual(listing.total_rate3, 1)
-        self.assertEqual(listing.total_rate4, 1)
-        self.assertEqual(listing.total_rate5, 1)
 
     def test_create_listing_minimal(self):
         # create a new listing with minimal data (title)
@@ -789,18 +286,6 @@ class ListingApiTest(APITestCase):
         response = self._request_helper(url, 'GET', username='wsmith', status_code=200)
         self.assertTrue(response.data.get('is_deleted'))
         self.assertEquals(self._validate_listing_map_keys(response.data), [])
-
-    def test_self_deleted_listing(self):
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/self/listing/'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-        titles = [i['id'] for i in response.data]
-        self.assertTrue('1' not in titles)
-        for listing_map in response.data:
-            self.assertEquals(self._validate_listing_map_keys(listing_map), [])
 
     def test_delete_listing_permission_denied(self):
         user = generic_model_access.get_profile('jones').user
@@ -1313,192 +798,6 @@ class ListingApiTest(APITestCase):
         # url = '/api/listing/%s/' % listing_id
         # response = self.client.get(url, data, format='json')
         # self.assertEqual(response.data['approval_status'], models.Listing.IN_PROGRESS)
-
-    def test_listing_activities(self):
-        action_log = []
-        # CREATED
-        url = '/api/listing/'
-
-        data = {
-            'title': 'mr jones app',
-            'security_marking': 'UNCLASSIFIED'
-        }
-
-        response = self._request_helper(url, 'POST', data=data, username='jones', status_code=201)
-        app_id = response.data['id']
-        data = response.data
-
-        # VERIFY that is was created
-        url = '/api/listing/{0!s}/activity/'.format(app_id)
-        response = self._request_helper(url, 'GET', username='jones', status_code=200)
-
-        activity_actions = [i['action'] for i in response.data]
-        self.assertEqual(len(activity_actions), 1)
-        action_log.insert(0, models.ListingActivity.CREATED)
-        self.assertEqual(activity_actions, action_log)
-        activity_agency = [i['listing']['agency'] for i in response.data]
-        self.assertEquals(json.dumps(activity_agency[0]), '{"title": "Ministry of Truth", "short_name": "Minitrue"}')
-
-        # MODIFIED
-        data['title'] = "mr jones mod app"
-        url = '/api/listing/{0!s}/'.format(app_id)
-        response = self._request_helper(url, 'PUT', data=data, username='jones', status_code=200)
-        data = response.data
-
-        # VERIFY that is was modified
-        url = '/api/listing/{0!s}/activity/'.format(app_id)
-        response = self._request_helper(url, 'GET', username='jones', status_code=200)
-
-        activity_actions = [i['action'] for i in response.data]
-        self.assertEqual(len(activity_actions), 2)
-        action_log.insert(0, models.ListingActivity.MODIFIED)
-        self.assertEqual(activity_actions, action_log)
-        activity_agency = [i['listing']['agency'] for i in response.data]
-        self.assertEquals(json.dumps(activity_agency[0]), '{"title": "Ministry of Truth", "short_name": "Minitrue"}')
-
-        # SUBMITTED
-        data['approval_status'] = models.Listing.PENDING
-        url = '/api/listing/{0!s}/'.format(app_id)
-        response = self._request_helper(url, 'PUT', data=data, username='jones', status_code=200)
-
-        # Verify that is was submitted
-        url = '/api/listing/{0!s}/activity/'.format(app_id)
-        response = self._request_helper(url, 'GET', username='jones', status_code=200)
-
-        activity_actions = [i['action'] for i in response.data]
-        self.assertEquals(len(activity_actions), 3)
-        action_log.insert(0, models.ListingActivity.SUBMITTED)
-        self.assertEqual(activity_actions, action_log)
-        self.assertTrue(models.ListingActivity.SUBMITTED in activity_actions)
-        activity_agency = [i['listing']['agency'] for i in response.data]
-        self.assertEquals(json.dumps(activity_agency[0]), '{"title": "Ministry of Truth", "short_name": "Minitrue"}')
-
-        # APPROVED_ORG
-
-        # APPROVED
-
-        # DISABLE
-        data['is_enabled'] = False
-        url = '/api/listing/{0!s}/'.format(app_id)
-        response = self._request_helper(url, 'PUT', data=data, username='jones', status_code=200)
-
-        # Verify that it was disabled
-        url = '/api/listing/{0!s}/activity/'.format(app_id)
-        response = self._request_helper(url, 'GET', username='jones', status_code=200)
-        activity_actions = [i['action'] for i in response.data]
-        self.assertEquals(len(activity_actions), 4)
-        action_log.insert(0, models.ListingActivity.DISABLED)
-        self.assertEqual(activity_actions, action_log)
-        activity_agency = [i['listing']['agency'] for i in response.data]
-        self.assertEquals(json.dumps(activity_agency[0]), '{"title": "Ministry of Truth", "short_name": "Minitrue"}')
-
-        # ENABLED
-        data['is_enabled'] = True
-        url = '/api/listing/{0!s}/'.format(app_id)
-        response = self._request_helper(url, 'PUT', data=data, username='jones', status_code=200)
-
-        # Verify that it was enabled
-        url = '/api/listing/{0!s}/activity/'.format(app_id)
-        response = self._request_helper(url, 'GET', username='jones', status_code=200)
-        activity_actions = [i['action'] for i in response.data]
-        self.assertEquals(len(activity_actions), 5)
-        action_log.insert(0, models.ListingActivity.ENABLED)
-        self.assertEqual(activity_actions, action_log)
-        activity_agency = [i['listing']['agency'] for i in response.data]
-        self.assertEquals(json.dumps(activity_agency[0]), '{"title": "Ministry of Truth", "short_name": "Minitrue"}')
-
-    def test_get_all_listing_activities(self):
-        """
-        All stewards should be able to access this endpoint (not std users)
-
-        Make sure org stewards of one org can't access activity for private
-        listings of another org.
-        """
-        expected_titles = ['Air Mail', 'Bread Basket', 'Chart Course',
-            'Chatter Box', 'Clipboard']
-        user = generic_model_access.get_profile('jones').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/activity/'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-        user = generic_model_access.get_profile('wsmith').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/activity/'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        titles = [i['listing']['title'] for i in response.data]
-        counter = 0
-        for i in expected_titles:
-            self.assertTrue(i in titles)
-            counter += 1
-        self.assertEqual(counter, len(expected_titles))
-
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/activity/'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        titles = [i['listing']['title'] for i in response.data]
-        titles = list(set(titles))
-        print('for bigbrother, got titles: {0!s}'.format(titles))
-        counter = 0
-        for i in expected_titles:
-            # Bread Basket is a private app, and bigbrother is not in that
-            # organization
-            if i != 'Bread Basket':
-                print('checking for app {0!s}'.format(i))
-                self.assertTrue(i in titles)
-                counter += 1
-        self.assertEqual(counter, len(expected_titles) - 1)
-
-    def test_get_self_listing_activities(self):
-        """
-        Returns activity for listings owned by current user
-        """
-        expected_titles = ['Bread Basket', 'Chatter Box']
-
-        user = generic_model_access.get_profile('julia').user
-        self.client.force_authenticate(user=user)
-        url = '/api/self/listings/activity/'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        titles = [i['listing']['title'] for i in response.data]
-        counter = 0
-        for i in expected_titles:
-            self.assertTrue(i in titles)
-            counter += 1
-        self.assertEqual(counter, len(expected_titles))
-
-    def test_get_listing_activities_offset_limit(self):
-        user = generic_model_access.get_profile('bigbrother').user
-        self.client.force_authenticate(user=user)
-        url = '/api/listings/activity/?offset=0&limit=24'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        data = response.data
-        self.assertTrue('count' in data)
-        self.assertTrue('previous' in data)
-        self.assertTrue('next' in data)
-        self.assertTrue('results' in data)
-        self.assertTrue('/api/listings/activity/?limit=24&offset=24' in data.get('next'))
-        self.assertEqual(data.get('previous'), None)
-        self.assertTrue(len(data.get('results')) >= 1)
-
-    def test_get_self_listing_activities_offset_limit(self):
-        user = generic_model_access.get_profile('aaronson').user
-        self.client.force_authenticate(user=user)
-        url = '/api/self/listings/activity/?offset=0&limit=24'
-        response = self.client.get(url, format='json')
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        data = response.data
-        self.assertTrue('count' in data)
-        self.assertTrue('previous' in data)
-        self.assertTrue('next' in data)
-        self.assertTrue('results' in data)
-        self.assertEqual(data.get('next'), None)
-        self.assertEqual(data.get('previous'), None)
-        self.assertTrue(len(data.get('results')) >= 0)
 
     def test_get_listings_with_query_params(self):
         """
